[ADD] Create RegisterForm component with validation logic, implement useRegister hook for registration handling, and update RegisterPage to utilize the new form structure
